feat:补充ir、wysiwyg封装到selection中
